requirements.txt
# Complete Requirements for Network Anomaly Detection System
# Consolidated from requirements.txt, requirements-dev.txt, and requirements-live.txt

# Web interface
flask>=2.3.0
gunicorn>=21.2.0

# Core data processing and ML libraries
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.11.0
joblib>=1.3.0

# Anomaly detection specific libraries
pyod>=1.1.0
imbalanced-learn>=0.11.0

# Advanced ML models
xgboost>=1.7.0
tensorflow>=2.13.0
keras>=2.13.0

# Visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0

# Feature Engineering
umap-learn>=0.5.3

# Real-time streaming
kafka-python>=2.0.2
pyspark>=3.4.0

# Packet capture libraries (for live system)
pyshark>=0.6.0
scapy>=2.5.0

# Configuration and utilities
PyYAML>=6.0
requests>=2.31.0
setuptools>=68.0.0
